refactor determine_display_status
Previously determine_display_status was a side-effecting function which mutated data classes to add two properties display_status and flag_resoved to classes. This made it very difficult to test and reason about.

This refactors in a couple of ways:

move flag_resolved into an is_flaggable variable to reflect its usage and pass to template where used
make determine_display_status a simpler function that returns a value and pass that to the templates
Adds tests for determine_display_status