Note: commented out few tests:
After the discussion between Ram and Navya, it was decided to temporarily comment out some tests. This is being done as we continue to refactor routes and macros in order to implement new flagging features. Once the new flagging features have been properly implemented, these tests will be uncommented and refactored accordingly.