FS-4082: show pagination uniformly on the scoring page
Uniformly show the sub-section pagination component on the "scoring"
page (either if the sub-section has already been scored or is it yet to
be scored).

This would navigate to the default sub-section page (the top link on the
left hand nav) to keep it consistent with the other pagination.

Note that this hasn't been confirmed as desired behaviour but was raised
as a question -- with a standard helper its no further effort to
implement so putting this in as a proof of concept.